* configure.ac: Rewrite strcasecmp to use AC_TRY_COMPILE instead 	of AC_CHECK_HEADER which keeps failing on FC5-x86. 	* libbase/container.cpp, container.h, jpeg.cpp, membuf.cpp, 	network.cpp, rc.cpp, triangulate_impl.h, tu_file.cpp, utility.cpp, 	zlib_adapter.cpp, axial_box.h, collision.cpp: Minor tweaks to 	remove compiler warnings with the latest GCC on FC5. Printing 	size_t should use %zd, and many unused parameters are commented 	out to reduce warnings. 	* libgeometry/kd_tree_dynamic.cpp, kd_tree_packed.cpp: Minor 	tweaks to remove compiler warnings with the latest GCC on 	FC5. Printing size_t should use %zd, and many unused parameters 	are commented out to reduce warnings. 	* macros/kde.m4: Use dnl to comment out serial, it's a reserved 	word now. 	* server/ActionExec.cpp, MovieClip.cpp, action_buffer.cpp, 	as_environment.h, button.cpp, dlist.cpp, font.cpp, fontlib.cpp, 	movie.h, movie_root.cpp, shape_character_def.cpp, /shm.cpp, 	sound.cpp, sprite_definition.cpp, sprite_instance.cpp, 	sprite_instance.h, xml.cpp, xmlnode.cpp, xmlsocket.cpp: Minor 	tweaks to remove compiler warnings with the latest GCC on 	FC5. Printing size_t should use %zd, and many unused parameters 	are commented out to reduce warnings. 	* server/asobj/NetConnection.cpp, NetConnection.h: Minor tweaks to 	remove compiler warnings with the latest GCC on FC5. Printing 	size_t should use %zd, and many unused parameters are commented 	out to reduce warnings. 	* server/swf/ASHandlers.cpp: Minor tweaks to remove compiler 	warnings with the latest GCC on FC5. Printing size_t should use 	%zd, and many unused parameters are commented out to reduce 	warnings. 	* utilities/parser.cpp:  Minor tweaks to remove compiler warnings 	with the latest GCC on FC5. Printing size_t should use %zd, and 	many unused parameters are commented out to reduce warnings. 	* plugin/klash/Makefile.am: Add libamf.

+++ server/sprite_definition.cpp
@@ -143,7 +143,7 @@ sprite_definition::read(stream* in)
 void
 sprite_definition::add_frame_name(const char* name)
 {
-	assert(m_loading_frame >= 0 && m_loading_frame < m_frame_count);
+	assert((int)m_loading_frame >= 0 && m_loading_frame < m_frame_count);
 
 	tu_string n = name;
 	size_t currently_assigned = 0;
